+++ src/utils.js
@@ -474,6 +474,8 @@ export const getIdentifiers = (identifiers = [], type, identifierTypesById) => {
     }
   });
 
+  if (!result.length) return null;
+
   return (
     <div>
       {result.map((value, index) => <div>{`${value}${index !== result.length - 1 ? ',' : ''}`}</div>)}
@@ -794,36 +796,6 @@ export const isMARCSource = (source) => {
 export const isUserInConsortiumMode = stripes => stripes.hasInterface('consortia');
 
 export const isInstanceShadowCopy = (source) => [`${CONSORTIUM_PREFIX}FOLIO`, `${CONSORTIUM_PREFIX}MARC`].includes(source);
-
-export const getUserTenantsPermissions = async (stripes, tenants = []) => {
-  const {
-    user: { user: { id } },
-    okapi: {
-      url,
-      token,
-    }
-  } = stripes;
-  const userTenantIds = tenants.map(tenant => tenant.id || tenant);
-
-  const promises = userTenantIds.map(async (tenantId) => {
-    const result = await fetch(`${url}/perms/users/${id}/permissions?full=true&indexField=userId`, {
-      headers: {
-        [OKAPI_TENANT_HEADER]: tenantId,
-        [CONTENT_TYPE_HEADER]: 'application/json',
-        ...(token && { [OKAPI_TOKEN_HEADER]: token }),
-      },
-      credentials: 'include',
-    });
-
-    const json = await result.json();
-
-    return { tenantId, ...json };
-  });
-
-  const userTenantsPermissions = await Promise.allSettled(promises);
-
-  return userTenantsPermissions.map(userTenantsPermission => userTenantsPermission.value);
-};
 
 export const hasMemberTenantPermission = (permissionName, tenantId, permissions = []) => {
   const tenantPermissions = permissions?.find(permission => permission?.tenantId === tenantId)?.permissionNames || [];