+++ src/withFacets.js
@@ -25,6 +25,7 @@ import {
   browseModeMap,
   browseCallNumberOptions,
   queryIndexes,
+  browseClassificationOptions,
 } from './constants';
 import { getAdvancedSearchTemplate } from './routes/buildManifestObject';
 
@@ -132,6 +133,10 @@ function withFacets(WrappedComponent) {
 
       if (queryIndex === browseModeOptions.SUBJECTS) {
         return 'search/subjects/facets';
+      }
+
+      if (Object.values(browseClassificationOptions).includes(queryIndex)) {
+        return 'search/classifications/facets';
       }
 
       return 'search/instances/facets';